fix(db): Allow string attributes to be null. Never coerce to ""
- Fixes a bug where reply_to_message_id = "" when we receive drafts from the sever.
- In general behavior should be consistent with API. Since they allow null we should too.

+++ src/flux/tasks/syncback-draft.coffee
@@ -90,12 +90,13 @@ class SyncbackDraftTask extends Task
     if draft.accountId is account.id
       return Promise.resolve(draft)
     else
-      NylasAPI.makeRequest
-        path: "/drafts/#{draft.serverId}"
-        accountId: draft.accountId
-        method: "DELETE"
-        body: {version: draft.version}
-        returnsModel: false
+      if draft.serverId
+        NylasAPI.makeRequest
+          path: "/drafts/#{draft.serverId}"
+          accountId: draft.accountId
+          method: "DELETE"
+          body: {version: draft.version}
+          returnsModel: false
 
       draft.accountId = account.id
       delete draft.serverId
